ajout de draw_matrix_multiply_axes_m4

+++ draw_make_matrix_rot_m4.c
@@ -6,17 +6,17 @@
 /*   By: snicolet <snicolet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/06/01 16:08:41 by snicolet          #+#    #+#             */
-/*   Updated: 2016/06/01 20:09:48 by snicolet         ###   ########.fr       */
+/*   Updated: 2016/06/01 20:28:43 by snicolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "draw.h"
 #include <math.h>
 
-t_m4	draw_make_matrix_m4_x(t_v4f pos, double rad, t_v4f scale)
+t_m4	draw_make_matrix_m4_x(t_v4f pos, float rad, t_v4f scale)
 {
-	const float		s = (float)sin(rad);
-	const float		c = (float)cos(rad);
+	const float		s = (float)sinf(rad);
+	const float		c = (float)cosf(rad);
 
 	return ((t_m4){
 		(t_v4f){scale.x, 0.0, 0.0, 0.0},
@@ -26,10 +26,10 @@ t_m4	draw_make_matrix_m4_x(t_v4f pos, double rad, t_v4f scale)
 	});
 }
 
-t_m4	draw_make_matrix_m4_y(t_v4f pos, double rad, t_v4f scale)
+t_m4	draw_make_matrix_m4_y(t_v4f pos, float rad, t_v4f scale)
 {
-	const float		s = (float)sin(rad);
-	const float		c = (float)cos(rad);
+	const float		s = (float)sinf(rad);
+	const float		c = (float)cosf(rad);
 
 	return ((t_m4){
 		(t_v4f){c, 0.0, -s * scale.y, 0.0},
@@ -39,10 +39,10 @@ t_m4	draw_make_matrix_m4_y(t_v4f pos, double rad, t_v4f scale)
 	});
 }
 
-t_m4	draw_make_matrix_m4_z(t_v4f pos, double rad, t_v4f scale)
+t_m4	draw_make_matrix_m4_z(t_v4f pos, float rad, t_v4f scale)
 {
-	const float		s = (float)sin(rad);
-	const float		c = (float)cos(rad);
+	const float		s = (float)sinf(rad);
+	const float		c = (float)cosf(rad);
 
 	return ((t_m4){
 		(t_v4f){c * scale.x, -s * scale.y, 0.0, 0.0},
